requirements.txt
pyttsx3>=2.90
PyPDF2>=3.0.1
python-docx>=1.1.2
ebooklib>=0.18
beautifulsoup4>=4.12.3
lxml>=5.3.0
keyboard>=0.13.5